[Bug 149] add user configuration support to ConfigurationLoader, made ReceiverConfiguration storable into Configuration
git-svn-id: http://svn.tryphon.org/kolaka/freecast/trunk@902 20e14386-5fff-0310-853f-b84fc43df78c

+++ source/org/kolaka/freecast/transport/receiver/PlaylistReceiver.java
@@ -53,7 +53,7 @@ public class PlaylistReceiver extends OggSourceFactoryReceiver {
     if (bandwidthControler instanceof StaticBandwidthControler) {
       long bandwidth = 
         ((StaticBandwidthControler) bandwidthControler).getBandwidth() / FileUtils.ONE_KB;
-      configuration.setBandwidth(bandwidth);
+      configuration.setBandwidth((int) bandwidth);
     }
     
     return configuration;
